requirements.txt
Django==1.9.3
Pillow==3.1.1
Unidecode==0.04.19
argparse==1.2.1
django-filer==1.1.1
django-filter==0.12.0
django-mptt==0.8.3
django-polymorphic==0.8.1
django-wysiwyg-redactor==0.4.9
easy-thumbnails==2.3
psycopg2==2.6.1
simplejson==3.8.2
wsgiref==0.1.2
